fix: resolve linter and vet issues in test files
- Rename errStringImpl/toolLoopErr/searchErr to xxxError format (errname)
- Remove duplicate package godoc comments (godoclint)
- Fix import grouping in cmd_tgtest_test.go (goimports)
- Remove unnecessary type conversions (unconvert)
- Add nolint for expected unicode test strings (gosmopolitan)
- Fix context.WithTimeout cancel leak in heartbeat test (vet)

+++ pkg/tui/cron_test.go
@@ -201,8 +201,8 @@ func TestRenderCronDetailComplete(t *testing.T) {
 	if m.cronService == nil {
 		t.Skip("cron service not initialized")
 	}
-	next := int64(time.Now().UnixMilli() + 3600000)
-	last := int64(time.Now().UnixMilli() - 60000)
+	next := time.Now().UnixMilli() + 3600000
+	last := time.Now().UnixMilli() - 60000
 	job, err := m.cronService.AddJobWithOptions("detail-job", cron.CronSchedule{Kind: "cron", Expr: "0 9 * * *", TZ: "UTC"}, "hello msg", true, "chan", "to", "session", "sess-key")
 	if err != nil {
 		t.Fatalf("add job: %v", err)
